Updated code so that some of the features of what happens each day now works.

+++ Source/RTU_Village/Player/RTUV_PlayerWidget.h
@@ -109,7 +109,7 @@ protected:
 	UButton* BtnDayReview;
 
 public:
-	void NewDayStart(int32 CurrentDay, int32 TotalDays, int32 TotalPeople, int32 AvailablePeople, float DefenceComplete, int32 TreesStored, int32 RawFoodStored, int32 CookedFoodStored, int32 HousesAvailable, int32 DefenceLastTurn, int32 TreesLastTurn, int32 HuntLastTurn, int32 CookLastTurn, int32 HousesLastTurn);
+	void NewDayStart(int32 CurrentDay, int32 TotalDays, int32 TotalPeople, int32 AvailablePeople, float DefenceComplete, float TreesStored, int32 RawFoodStored, int32 CookedFoodStored, int32 HousesAvailable, int32 DefenceLastTurn, int32 TreesLastTurn, int32 HuntLastTurn, int32 CookLastTurn, int32 HousesLastTurn);
 	void OnDayEnd(const FText& TextIn);
 private:
 	UPROPERTY()
